modules/j1939-sensor/requirements.txt
viam-sdk>=0.25.0
python-can>=4.0
